Pull out resuable methods into utils
